+++ test/cases/option-preload-dynamic-import-filter-fn/src/noPreload.js
@@ -1,0 +1,1 @@
+console.log('>> noPreload.js');